requirements.txt
streamlit
requests
pandas
numpy
plotly
scipy
python-dotenv